LPD-8519 Merge SharepointRepositoryFactoryProvider with SharepointRepositoryDefiner
